--- packages/ui-contexts/src/ServerContext/methods/saveSettings.ts
@@ -1,9 +1,0 @@
-import type { ISetting } from '@rocket.chat/core-typings';
-
-type SettingChange = {
-	_id: ISetting['_id'];
-	value: unknown;
-	editor?: unknown;
-};
-
-export type SaveSettingsMethod = (changes: SettingChange[]) => true;
