--- packages/ui-contexts/src/ServerContext/methods/saveUserPreferences.ts
@@ -1,36 +1,0 @@
-type UserPreferences = {
-	language: string;
-	newRoomNotification: string;
-	newMessageNotification: string;
-	clockMode: number;
-	useEmojis: boolean;
-	convertAsciiEmoji: boolean;
-	saveMobileBandwidth: boolean;
-	collapseMediaByDefault: boolean;
-	autoImageLoad: boolean;
-	emailNotificationMode: string;
-	unreadAlert: boolean;
-	notificationsSoundVolume: number;
-	desktopNotifications: string;
-	pushNotifications: string;
-	enableAutoAway: boolean;
-	highlights: string[];
-	hideUsernames: boolean;
-	hideRoles: boolean;
-	displayAvatars: boolean;
-	hideFlexTab: boolean;
-	sendOnEnter: string;
-	idleTimeLimit: number;
-	sidebarShowFavorites: boolean;
-	sidebarShowUnread: boolean;
-	sidebarSortby: string;
-	sidebarViewMode: string;
-	sidebarDisplayAvatar: boolean;
-	sidebarGroupByType: boolean;
-	muteFocusedConversations: boolean;
-	dontAskAgainList: { action: string; label: string }[];
-	themeAppearence: 'auto' | 'light' | 'dark';
-	receiveLoginDetectionEmail: boolean;
-};
-
-export type SaveUserPreferencesMethod = (preferences: Partial<UserPreferences>) => boolean;
